+++ backend/src/battles/dto/battle-join-request.dto.ts
@@ -1,0 +1,16 @@
+import { IsNotEmpty, IsOptional, IsString } from 'class-validator'
+import type { BattleTeam } from '../types/battles.types'
+
+export class BattleJoinRequestDto {
+  @IsString()
+  @IsNotEmpty()
+  battleId: string
+
+  @IsString()
+  @IsNotEmpty()
+  team: BattleTeam
+
+  @IsString()
+  @IsOptional()
+  password?: string
+}
