--- backend/src/battles/dto/join-battle.dto.ts
@@ -1,5 +1,0 @@
-export class JoinBattleDto {
-  battleId: string
-  team: string
-  password?: string
-}
